requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
jinja2
python-dotenv
passlib